_scripts/requirements.txt
pyyaml >= 5.4.1, < 6
